refactor: add Javadocs, refactor code, improve tests, and fix Checkstyle issues for factories
